modifying

+++ 0x1E-search_algorithms/0-linear.c
@@ -16,7 +16,7 @@ int linear_search(int *array, size_t size, int value)
 	{
 		printf("Value checked array[%lu] = [%d]\n", a, array[a]);
 		if (array[a] == value)
-			return ((int) a);
+			return (a);
 
 	}
 	return (-1);
